requirements.txt
certifi==2018.8.13
chardet==3.0.4
Django==2.1
django-dotenv==1.4.2
djangorestframework==3.8.2
idna==2.7
pytz==2018.5
requests==2.19.1
stripe==2.5.0
urllib3==1.23
